requirements.txt
Django==5.2.8
djangorestframework==3.16.1
Pillow==12.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.8.2
